LayoutTests/storage/indexeddb/modern/opendatabase-request-expected.txt

This test calls open on window.indexedDB in various ways, verifying that IDBOpenDBRequest objects are returned when expected and exceptions are thrown when expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: [object IDBOpenDBRequest]
ALERT: [object IDBOpenDBRequest]
ALERT: TypeError: Not enough arguments
ALERT: TypeError: Type error
PASS successfullyParsed is true

TEST COMPLETE
